Split scheduler unit tests.
